requirements.txt
# Airflow dependencies
apache-airflow==2.3.0
apache-airflow-providers-postgres==2.1.0
apache-airflow-providers-redis==2.0.0

# Database driver
psycopg2-binary==2.9.1

# Data processing libraries
requests==2.25.1
pandas==1.3.0
pyarrow==5.0.0
